+++ backend/src/dashboard/router.py
@@ -1,0 +1,2 @@
+# Re-expose router object.
+from dashboard.service import router
